Improve reliability of `clean-repo-tabs` and `more-dropdown-links`
